move leaderboard notifications in different package
