+++ notifications/src/main/java/com/dojo/notifications/service/LeaderboardNotifierService.java
@@ -4,8 +4,8 @@ import com.dojo.notifications.model.contest.Contest;
 import com.dojo.notifications.model.contest.enums.EventType;
 import com.dojo.notifications.model.contest.enums.NotifierType;
 import com.dojo.notifications.model.leaderboard.Leaderboard;
-import com.dojo.notifications.model.notification.CommonLeaderboardNotification;
-import com.dojo.notifications.model.notification.PersonalLeaderboardNotification;
+import com.dojo.notifications.model.notification.leaderboard.CommonLeaderboardNotification;
+import com.dojo.notifications.model.notification.leaderboard.PersonalLeaderboardNotification;
 import com.dojo.notifications.model.request.SelectRequest;
 import com.dojo.notifications.model.user.Participant;
 import com.dojo.notifications.model.user.UserDetails;
